Migrated DragHandle to ts
